Refactoring code according to guidelines and removing unnecassary packages and files

+++ src/main/java/io/mountblue/redditclone/utils/requests/CreatePostRequest.java
@@ -17,5 +17,5 @@ public class CreatePostRequest {
 
     private boolean isDraft;
 
-    private int subredditId;
+    private Long subredditId;
 }
